requirements.txt
tkinterdnd2>=0.3.0
Pillow>=10.0.1
openpyxl>=3.1.0
requests>=2.31.0
pandas>=2.0.0